Split wasm and rest bergamot api client code into separate classes
